mediasoup-kurento-filter: Add support for SRTP with mediasoup
mediasoup 3.5 now includes support for SRTP, with a new class named
PlainTransport. This change adds all required code to the demo app, to
leave it in a working state with RTP and expected working state with
SRTP.
Only remaining change is updating mediasoup and moving to use of the new
PlainTransport class.