mediasoup-kurento-filter: use mediasoup "announcedIp" for remote KMS
When KMS is not running in localhost, the config must be changed to
reflect what is the correct IP address where KMS can send data to
mediasoup.

For example if KMS runs from within a Docker container, from its point
of view this demo will be running at the IP address "172.17.0.1".