mediasoup-kurento-filter: add "logging" module using "debug" package, remove console
This allows to set a desired log level and have dynamic logs that are
printed or omitted depending on what was the log level set by the user.
Use the DEBUG environment variable to set different log levels.

The console was getting garbled data because of colorized output, and
also it wasn't that useful because for this demo the node server is
typically running on its own window so there wasn't much value in having
server logs in the UI.

Also rename some variables to shorter names:
* "capabilities" -> "caps"
* "parameters" -> "params"
to keep the same naming than in other projects we're working on.